+++ karma.conf.js
@@ -13,7 +13,6 @@ module.exports = function (config) {
 
     // list of files / patterns to load in the browser
     files: [
-      'node_modules/babel-polyfill/dist/polyfill.js',
       'dist/zetapush.js',
       'test/**/*.spec.js'
     ],