+++ src/client/basic.js
@@ -203,6 +203,29 @@ export class Client {
   getUserId() {
     return this.helper.getUserId()
   }
+  /*
+   * Get the client user info
+   * @return {Object}
+   * @example
+   * // Create new ZetaPush Client
+   * const client = new Client({
+   *   sandboxId: '<YOUR-SANDBOX-ID>',
+   *   authentication: () => Authentication.simple({
+   *     login: '<YOUR-USER-LOGIN>',
+   *     password: '<YOUR-USER-PASSWORD>'
+   *   })
+   * })
+   * // Add connection establised listener
+   * client.onConnectionEstablished(() => {
+   *   console.log('onConnectionEstablished')
+   *   const profile = client.getUserInfo()
+   *   console.log('Your profile', profile)
+   * })
+   * client.connect()
+   */
+  getUserInfo() {
+    return this.helper.getUserInfo()
+  }
   /**
    * Remove a connection status listener
    * @param {number} handler