+++ test/prerequisites.spec.js
@@ -1,8 +1,8 @@
-describe('Prerequisites', function () {
-  it('Should have Promise global function', function () {
+describe('Prerequisites', () => {
+  it('Should have Promise global function', () => {
     expect(typeof Promise).toBe('function')
   })
-  it('Should have fetch global function', function () {
+  it('Should have fetch global function', () => {
     expect(typeof fetch).toBe('function')
   })
 })